requirements.txt
dateparser==0.7.1
georinex==1.10.0
kiwisolver==1.1.0
numpy==1.16.2
pandas==0.24.2
Pympler==0.7
pyparsing==2.4.0
pyproj==2.1.3
pyshp==2.1.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
regex==2019.4.14
scipy==1.2.1
six==1.12.0
tzlocal==1.5.1
xarray==0.12.1
